+++ settings.py
@@ -43,7 +43,13 @@ INSTALLED_APPS = [
     'reference_data',
     'matchmaker',
     'social_django',
+    'panelapp',
 ]
+
+###########
+# PanelApp
+###########
+PANEL_APP_API_URL = os.environ.get('PANEL_APP_API_URL', 'https://panelapp.url/api/v1')
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',