+++ ui/shared/components/panel/variants/Annotations.jsx
@@ -163,8 +163,8 @@ const BaseSearchLinks = React.memo(({ variant, mainTranscript, genesById }) => {
   }
   links.unshift(
     <Popup
-      key="seqr"
-      trigger={<SearchResultsLink buttonText="seqr" {...seqrLinkProps} />}
+      key="seqr-search"
+      trigger={<SearchResultsLink key="seqr" buttonText="seqr" {...seqrLinkProps} />}
       content="Search for this variant across all your seqr projects"
       size="tiny"
     />,
@@ -279,7 +279,7 @@ const Annotations = React.memo(({ variant }) => {
         </Modal>
       }
       {svType && end && <b><HorizontalSpacer width={5} />{svSizeDisplay(end - pos)}</b>}
-      {numExon && <b>, {numExon} exons</b>}
+      {Number.isInteger(numExon) && <b>, {numExon} exons</b>}
       { lofDetails &&
         <span>
           <HorizontalSpacer width={12} />