+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # _seqr_ Changes
 
 ## dev
+* Change family note fields into lists of notes (REQUIRES DB MIGRATION)
+
+## 9/10/21
+* Use google storage API instead of gsutil for IGV 
 
 ## 9/3/21
 * Add last updated information for family analysis_status (REQUIRES DB MIGRATION)